Transfer signal function to signals.py module
