+++ python/apps/vacancies/signals.py
@@ -1,0 +1,9 @@
+from apps.requests.models import Request
+
+
+def vacancy_created(sender, **kwargs):
+    if kwargs['created']:
+        instance = Request.objects.get(pk=kwargs['instance'].request.id)
+        instance.is_vacancy_created = True
+        instance.save()
+
